refactor(data-grid): split 1455-line shell into 15 focused modules + fix table-node
Breaks data-grid.tsx (1455 → 245 lines) into a focused shell plus 15 sibling
files under data-grid/. Behavior is unchanged; ownership is now clear and
each piece is independently testable.

Hooks (one concern each):
- use-cell-editing, use-cell-selection, use-row-selection
- use-grid-keyboard, use-column-resize, use-focused-cell
- use-context-menu-reporting, use-right-drag-scroll

Child components:
- cell-value, draft-row, empty-states, grid-body, grid-header

Pure modules:
- selection.ts (rectangle helpers), types.ts (CellPosition, ContextMenuState)

Also restores the broken imports in
schema-visualizer/components/table-node.tsx (Handle, Position, NodeProps,
Key, LinkIcon) — tsc -b is now clean across the workspace.

Bumps the version to 0.0.109 to bundle this refactor with the AI key
store work already on the branch into a single release.
