+++ apps/desktop/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@dora/desktop",
-	"version": "0.0.108",
+	"version": "0.0.109",
 	"private": true,
 	"type": "module",
 	"scripts": {
