+++ apps/desktop/src-tauri/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "app"
-version = "0.0.108"
+version = "0.0.109"
 description = ""
 authors = ["Remco Stoeten"]
 license = "GPLv3"
