+++ apps/desktop/src-tauri/tauri.conf.json
@@ -1,7 +1,7 @@
 {
 	"$schema": "../node_modules/@tauri-apps/cli/config.schema.json",
 	"productName": "Dora",
-	"version": "0.0.108",
+	"version": "0.0.109",
 	"identifier": "com.Dora.dev",
 	"build": {
 		"frontendDist": "../dist",
