+++ apps/desktop/src/features/database-studio/components/data-grid/types.ts
@@ -1,0 +1,16 @@
+export type EditingCell = {
+	rowIndex: number
+	columnName: string
+}
+
+export type CellPosition = {
+	row: number
+	col: number
+}
+
+export type ContextMenuState = {
+	kind: 'cell' | 'row'
+	cell: { row: number; col: number }
+	x: number
+	y: number
+} | null
